+++ drum-machine/src/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import DrumMachine from './App';
+
+ReactDOM.render(
+  <DrumMachine />,
+  document.getElementById('root')
+);
