--- drum-machine/style.css
@@ -1,209 +1,0 @@
-* {
-  margin: 0;
-  box-sizing: border-box;
-  background-color: black;
-  position: relative;
-  outline: none;
-}
-:root {
-  --thumb-background: url('https://i.postimg.cc/xCYYnr8m/Volume-Knob-Light.png');
-}
-html {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-#drum-machine {
-  display: inline-block;
-}
-.no-drag {
-  user-drag: none;
-  -webkit-user-drag: none;
-  user-select: none;
-  -moz-user-select: none;
-  -webkit-user-select: none;
-  -ms-user-select: none;
-}
-img {
-  background: none;
-}
-.power-button {
-  position: absolute;
-  top: -365px;
-  left: 143px;
-  height: 37px;
-  width: 35px;
-  overflow: hidden;
-  border-radius: 50%;
-  object-fit: cover;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.power-button:hover {
-  cursor: pointer;
-  opacity: 90%;
-}
-#bank-1{
-  position: absolute;
-  width: 50px;
-  object-fit: none;
-  top: 49px;
-  left: 197px;
-  background: none;
-  display: flex;
-}
-#bank-2 {
-  position: absolute;
-  top: 49px;
-  left: 244px;
-  width: 50px;
-  object-fit: none;
-  background: none;
-  display: flex;
-}
-#screen {
-  position: absolute;
-  top: -392px;
-  right: 20px
-}
-#display {
-  position: absolute;
-  top: 49px;
-  left: 345px;
-  background: none;
-  color: #082e5e;
-  font-family: arial;
-  user-select: none;
-}
-#track-light {
-  position: absolute;
-  top: 148px;
-  right: 87px;
-}
-#volume-name {
-  position: absolute;
-  top: 130px;
-  right: 31px;
-}
-.slider-container {
-  position: absolute;
-  transform: rotate(-90deg);
-  width: 200px;
-  top: 250px;
-  right: -19.5px;
-  background: none;
-}
-.volume-slider {
-  appearance: none;
-  background: none;
-  width: 213px;
-  height: 1px;
-}
-.volume-slider::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  background-image: var(--thumb-background);
-  height: 40px;
-  width: 40px;
-  position: relative;
-  top: -7px;
-  transform: rotate(90deg);
-}
-.volume-slider::-webkit-slider-thumb:hover {
-  cursor: grab;
-}
-.drum-pad {
-  position: absolute;
-}
-#drum-pad1 {
-  top: 144px;
-  left: 59px;
-}
-#drum-pad2 {
-  top: 144px;
-  left: 194px;
-}
-#drum-pad3 {
-  top: 144px;
-  left: 329px;
-}
-#drum-pad4 {
-  top: 224px;
-  left: 59px;
-}
-#drum-pad5 {
-  top: 224px;
-  left: 194px;
-}
-#drum-pad6 {
-  top: 224px;
-  left: 329px;
-}
-#drum-pad7 {
-  top: 304px;
-  left: 59px;
-}
-#drum-pad8 {
-  top: 304px;
-  left: 194px;
-}
-#drum-pad9 {
-  top: 304px;
-  left: 329px;
-}
-div:focus {
-  outline: none;
-}
-p {
-  position: absolute;
-  color: #ededed;
-  font-family: arial;
-  background: none;
-}
-.q {
-  top: 167px;
-  left: 106px;
-  user-select: none;
-}
-.w {
-  top: 167px;
-  left: 241px;
-  user-select: none;
-}
-.e {
-  top: 167px;
-  left: 376px;
-  user-select: none;
-}
-.a {
-  top: 247px;
-  left: 106px;
-  user-select: none;
-}
-.s {
-  top: 247px;
-  left: 241px;
-  user-select: none;
-}
-.d {
-  top: 247px;
-  left: 376px;
-  user-select: none;
-}
-.z {
-  top: 327px;
-  left: 106px;
-  user-select: none;
-}
-.x {
-  top: 327px;
-  left: 241px;
-  user-select: none;
-}
-.c {
-  top: 327px;
-  left: 376px;
-  user-select: none;
-}
